+++ packages/audio-station/src/util.ts
@@ -1,0 +1,1 @@
+export type CustomEventHandler<T> = (event: CustomEvent<T>) => void;